+++ benchmarks/sha2.cc
@@ -6,7 +6,7 @@
  *    - http://opensource.org/licenses/MIT
  */
 #include "EverCrypt_Hash.h"
-#include "Hacl_Streaming_SHA2.h"
+#include "Hacl_Hash_SHA2.h"
 
 #ifdef HACL_CAN_COMPILE_VEC128
 #include "Hacl_Hash_Blake2s_128.h"
@@ -159,7 +159,7 @@ BENCHMARK_CAPTURE(HACL_Sha2_oneshot,
                   sha2_224,
                   HACL_HASH_SHA2_224_DIGEST_LENGTH,
                   expected_digest_sha2_224,
-                  Hacl_Hash_SHA2_hash_224)
+                  Hacl_Streaming_SHA2_hash_224)
   ->Setup(DoSetup);
 
 BENCHMARK_CAPTURE(EverCrypt_Sha2_oneshot,
@@ -182,7 +182,7 @@ BENCHMARK_CAPTURE(HACL_Sha2_oneshot,
                   sha2_256,
                   HACL_HASH_SHA2_256_DIGEST_LENGTH,
                   expected_digest_sha2_256,
-                  Hacl_Hash_SHA2_hash_256)
+                  Hacl_Streaming_SHA2_hash_256)
   ->Setup(DoSetup);
 
 BENCHMARK_CAPTURE(EverCrypt_Sha2_oneshot,
@@ -205,7 +205,7 @@ BENCHMARK_CAPTURE(HACL_Sha2_oneshot,
                   sha2_384,
                   HACL_HASH_SHA2_384_DIGEST_LENGTH,
                   expected_digest_sha2_384,
-                  Hacl_Hash_SHA2_hash_384)
+                  Hacl_Streaming_SHA2_hash_384)
   ->Setup(DoSetup);
 
 BENCHMARK_CAPTURE(EverCrypt_Sha2_oneshot,
@@ -228,7 +228,7 @@ BENCHMARK_CAPTURE(HACL_Sha2_oneshot,
                   sha2_512,
                   HACL_HASH_SHA2_512_DIGEST_LENGTH,
                   expected_digest_sha2_512,
-                  Hacl_Hash_SHA2_hash_512)
+                  Hacl_Streaming_SHA2_hash_512)
   ->Setup(DoSetup);
 
 BENCHMARK_CAPTURE(EverCrypt_Sha2_oneshot,
