+++ config/default_config.cmake
@@ -9,7 +9,6 @@ set(SOURCES_std
 	${PROJECT_SOURCE_DIR}/src/Hacl_Hash_Blake2.c
 	${PROJECT_SOURCE_DIR}/src/Lib_Memzero0.c
 	${PROJECT_SOURCE_DIR}/src/Hacl_Ed25519.c
-	${PROJECT_SOURCE_DIR}/src/Hacl_Streaming_SHA2.c
 	${PROJECT_SOURCE_DIR}/src/Hacl_EC_Ed25519.c
 	${PROJECT_SOURCE_DIR}/src/Hacl_Hash_Base.c
 	${PROJECT_SOURCE_DIR}/src/Hacl_Streaming_Blake2.c
@@ -127,16 +126,13 @@ set(INCLUDES
 	${PROJECT_SOURCE_DIR}/include/Hacl_Hash_Blake2.h
 	${PROJECT_SOURCE_DIR}/include/Lib_Memzero0.h
 	${PROJECT_SOURCE_DIR}/include/internal/Hacl_Ed25519.h
-	${PROJECT_SOURCE_DIR}/include/internal/Hacl_Streaming_SHA2.h
 	${PROJECT_SOURCE_DIR}/include/internal/Hacl_SHA2_Generic.h
-	${PROJECT_SOURCE_DIR}/include/internal/../Hacl_Streaming_SHA2.h
 	${PROJECT_SOURCE_DIR}/include/Hacl_Streaming_Types.h
 	${PROJECT_SOURCE_DIR}/include/internal/Hacl_Ed25519_PrecompTable.h
 	${PROJECT_SOURCE_DIR}/include/internal/Hacl_Curve25519_51.h
 	${PROJECT_SOURCE_DIR}/include/internal/Hacl_Bignum25519_51.h
 	${PROJECT_SOURCE_DIR}/include/internal/../Hacl_Curve25519_51.h
 	${PROJECT_SOURCE_DIR}/include/internal/../Hacl_Ed25519.h
-	${PROJECT_SOURCE_DIR}/include/Hacl_Streaming_SHA2.h
 	${PROJECT_SOURCE_DIR}/include/Hacl_EC_Ed25519.h
 	${PROJECT_SOURCE_DIR}/include/Hacl_Hash_Base.h
 	${PROJECT_SOURCE_DIR}/include/internal/Hacl_Hash_Blake2.h
@@ -270,11 +266,9 @@ set(PUBLIC_INCLUDES
 	${PROJECT_SOURCE_DIR}/include/Hacl_Hash_SHA2.h
 	${PROJECT_SOURCE_DIR}/include/Hacl_Hash_Blake2.h
 	${PROJECT_SOURCE_DIR}/include/Lib_Memzero0.h
-	${PROJECT_SOURCE_DIR}/include/internal/../Hacl_Streaming_SHA2.h
 	${PROJECT_SOURCE_DIR}/include/Hacl_Streaming_Types.h
 	${PROJECT_SOURCE_DIR}/include/internal/../Hacl_Curve25519_51.h
 	${PROJECT_SOURCE_DIR}/include/internal/../Hacl_Ed25519.h
-	${PROJECT_SOURCE_DIR}/include/Hacl_Streaming_SHA2.h
 	${PROJECT_SOURCE_DIR}/include/Hacl_EC_Ed25519.h
 	${PROJECT_SOURCE_DIR}/include/Hacl_Hash_Base.h
 	${PROJECT_SOURCE_DIR}/include/internal/../Hacl_Hash_Blake2.h
