+++ include/Hacl_Ed25519.h
@@ -36,7 +36,6 @@ extern "C" {
 #include "krml/internal/target.h"
 
 #include "Hacl_Streaming_Types.h"
-#include "Hacl_Streaming_SHA2.h"
 #include "Hacl_Krmllib.h"
 #include "Hacl_Hash_SHA2.h"
