+++ include/Hacl_Streaming_Blake2.h
@@ -35,6 +35,7 @@ extern "C" {
 #include "krml/lowstar_endianness.h"
 #include "krml/internal/target.h"
 
+#include "Hacl_Streaming_Types.h"
 #include "Hacl_Krmllib.h"
 #include "Hacl_Hash_Blake2.h"
 
@@ -82,7 +83,7 @@ void Hacl_Streaming_Blake2_blake2s_32_no_key_init(Hacl_Streaming_Blake2_blake2s_
 /**
   Update function when there is no key; 0 = success, 1 = max length exceeded
 */
-uint32_t
+Hacl_Streaming_Types_error_code
 Hacl_Streaming_Blake2_blake2s_32_no_key_update(
   Hacl_Streaming_Blake2_blake2s_32_state *p,
   uint8_t *data,
@@ -117,7 +118,7 @@ void Hacl_Streaming_Blake2_blake2b_32_no_key_init(Hacl_Streaming_Blake2_blake2b_
 /**
   Update function when there is no key; 0 = success, 1 = max length exceeded
 */
-uint32_t
+Hacl_Streaming_Types_error_code
 Hacl_Streaming_Blake2_blake2b_32_no_key_update(
   Hacl_Streaming_Blake2_blake2b_32_state *p,
   uint8_t *data,
