+++ ocaml/lib/EverCrypt_Hash_bindings.ml
@@ -30,10 +30,6 @@ module Bindings(F:Cstubs.FOREIGN) =
     let everCrypt_Hash_update_multi_256 =
       foreign "EverCrypt_Hash_update_multi_256"
         ((ptr uint32_t) @-> (ocaml_bytes @-> (uint32_t @-> (returning void))))
-    let everCrypt_Hash_update_last_256 =
-      foreign "EverCrypt_Hash_update_last_256"
-        ((ptr uint32_t) @->
-           (uint64_t @-> (ocaml_bytes @-> (uint32_t @-> (returning void)))))
     let everCrypt_Hash_Incremental_hash_len =
       foreign "EverCrypt_Hash_Incremental_hash_len"
         (spec_Hash_Definitions_hash_alg @-> (returning uint32_t))
