+++ ocaml/lib/Hacl_Streaming_Blake2_bindings.ml
@@ -2,6 +2,9 @@ open Ctypes
 module Bindings(F:Cstubs.FOREIGN) =
   struct
     open F
+    module Hacl_Streaming_Types_applied =
+      (Hacl_Streaming_Types_bindings.Bindings)(Hacl_Streaming_Types_stubs)
+    open Hacl_Streaming_Types_applied
     type hacl_Streaming_Blake2_blake2s_32_block_state =
       [ `hacl_Streaming_Blake2_blake2s_32_block_state ] structure
     let (hacl_Streaming_Blake2_blake2s_32_block_state :
@@ -57,7 +60,8 @@ module Bindings(F:Cstubs.FOREIGN) =
     let hacl_Streaming_Blake2_blake2s_32_no_key_update =
       foreign "Hacl_Streaming_Blake2_blake2s_32_no_key_update"
         ((ptr hacl_Streaming_Blake2_blake2s_32_state) @->
-           (ocaml_bytes @-> (uint32_t @-> (returning uint32_t))))
+           (ocaml_bytes @->
+              (uint32_t @-> (returning hacl_Streaming_Types_error_code))))
     let hacl_Streaming_Blake2_blake2s_32_no_key_finish =
       foreign "Hacl_Streaming_Blake2_blake2s_32_no_key_finish"
         ((ptr hacl_Streaming_Blake2_blake2s_32_state) @->
@@ -74,7 +78,8 @@ module Bindings(F:Cstubs.FOREIGN) =
     let hacl_Streaming_Blake2_blake2b_32_no_key_update =
       foreign "Hacl_Streaming_Blake2_blake2b_32_no_key_update"
         ((ptr hacl_Streaming_Blake2_blake2b_32_state) @->
-           (ocaml_bytes @-> (uint32_t @-> (returning uint32_t))))
+           (ocaml_bytes @->
+              (uint32_t @-> (returning hacl_Streaming_Types_error_code))))
     let hacl_Streaming_Blake2_blake2b_32_no_key_finish =
       foreign "Hacl_Streaming_Blake2_blake2b_32_no_key_finish"
         ((ptr hacl_Streaming_Blake2_blake2b_32_state) @->
