+++ src/Hacl_Ed25519.c
@@ -25,8 +25,8 @@
 
 #include "internal/Hacl_Ed25519.h"
 
-#include "internal/Hacl_Streaming_SHA2.h"
 #include "internal/Hacl_Krmllib.h"
+#include "internal/Hacl_Hash_SHA2.h"
 #include "internal/Hacl_Ed25519_PrecompTable.h"
 #include "internal/Hacl_Curve25519_51.h"
 #include "internal/Hacl_Bignum_Base.h"
@@ -1716,8 +1716,9 @@ static inline void sha512_pre_msg(uint8_t *hash, uint8_t *prefix, uint32_t len, 
   Hacl_Streaming_MD_state_64 p = s;
   Hacl_SHA2_Scalar32_sha512_init(block_state);
   Hacl_Streaming_MD_state_64 *st = &p;
-  uint32_t uu____0 = Hacl_Streaming_SHA2_update_512(st, prefix, (uint32_t)32U);
-  uint32_t uu____1 = Hacl_Streaming_SHA2_update_512(st, input, len);
+  Hacl_Streaming_Types_error_code
+  uu____0 = Hacl_Streaming_SHA2_update_512(st, prefix, (uint32_t)32U);
+  Hacl_Streaming_Types_error_code uu____1 = Hacl_Streaming_SHA2_update_512(st, input, len);
   Hacl_Streaming_SHA2_finish_512(st, hash);
 }
 
@@ -1737,9 +1738,11 @@ sha512_pre_pre2_msg(
   Hacl_Streaming_MD_state_64 p = s;
   Hacl_SHA2_Scalar32_sha512_init(block_state);
   Hacl_Streaming_MD_state_64 *st = &p;
-  uint32_t uu____0 = Hacl_Streaming_SHA2_update_512(st, prefix, (uint32_t)32U);
-  uint32_t uu____1 = Hacl_Streaming_SHA2_update_512(st, prefix2, (uint32_t)32U);
-  uint32_t uu____2 = Hacl_Streaming_SHA2_update_512(st, input, len);
+  Hacl_Streaming_Types_error_code
+  uu____0 = Hacl_Streaming_SHA2_update_512(st, prefix, (uint32_t)32U);
+  Hacl_Streaming_Types_error_code
+  uu____1 = Hacl_Streaming_SHA2_update_512(st, prefix2, (uint32_t)32U);
+  Hacl_Streaming_Types_error_code uu____2 = Hacl_Streaming_SHA2_update_512(st, input, len);
   Hacl_Streaming_SHA2_finish_512(st, hash);
 }
 
@@ -1778,7 +1781,7 @@ static inline void point_mul_g_compress(uint8_t *out, uint8_t *s)
 
 static inline void secret_expand(uint8_t *expanded, uint8_t *secret)
 {
-  Hacl_Hash_SHA2_hash_512(secret, (uint32_t)32U, expanded);
+  Hacl_Streaming_SHA2_hash_512(secret, (uint32_t)32U, expanded);
   uint8_t *h_low = expanded;
   uint8_t h_low0 = h_low[0U];
   uint8_t h_low31 = h_low[31U];
