+++ src/Hacl_P256.c
@@ -1563,7 +1563,7 @@ Hacl_P256_ecdsa_sign_p256_sha2(
 {
   uint64_t m_q[4U] = { 0U };
   uint8_t mHash[32U] = { 0U };
-  Hacl_Hash_SHA2_hash_256(msg, msg_len, mHash);
+  Hacl_Streaming_SHA2_hash_256(msg, msg_len, mHash);
   uint8_t *mHash32 = mHash;
   bn_from_bytes_be4(m_q, mHash32);
   qmod_short(m_q, m_q);
@@ -1595,7 +1595,7 @@ Hacl_P256_ecdsa_sign_p256_sha384(
 {
   uint64_t m_q[4U] = { 0U };
   uint8_t mHash[48U] = { 0U };
-  Hacl_Hash_SHA2_hash_384(msg, msg_len, mHash);
+  Hacl_Streaming_SHA2_hash_384(msg, msg_len, mHash);
   uint8_t *mHash32 = mHash;
   bn_from_bytes_be4(m_q, mHash32);
   qmod_short(m_q, m_q);
@@ -1627,7 +1627,7 @@ Hacl_P256_ecdsa_sign_p256_sha512(
 {
   uint64_t m_q[4U] = { 0U };
   uint8_t mHash[64U] = { 0U };
-  Hacl_Hash_SHA2_hash_512(msg, msg_len, mHash);
+  Hacl_Streaming_SHA2_hash_512(msg, msg_len, mHash);
   uint8_t *mHash32 = mHash;
   bn_from_bytes_be4(m_q, mHash32);
   qmod_short(m_q, m_q);
@@ -1704,7 +1704,7 @@ Hacl_P256_ecdsa_verif_p256_sha2(
 {
   uint64_t m_q[4U] = { 0U };
   uint8_t mHash[32U] = { 0U };
-  Hacl_Hash_SHA2_hash_256(msg, msg_len, mHash);
+  Hacl_Streaming_SHA2_hash_256(msg, msg_len, mHash);
   uint8_t *mHash32 = mHash;
   bn_from_bytes_be4(m_q, mHash32);
   qmod_short(m_q, m_q);
@@ -1734,7 +1734,7 @@ Hacl_P256_ecdsa_verif_p256_sha384(
 {
   uint64_t m_q[4U] = { 0U };
   uint8_t mHash[48U] = { 0U };
-  Hacl_Hash_SHA2_hash_384(msg, msg_len, mHash);
+  Hacl_Streaming_SHA2_hash_384(msg, msg_len, mHash);
   uint8_t *mHash32 = mHash;
   bn_from_bytes_be4(m_q, mHash32);
   qmod_short(m_q, m_q);
@@ -1764,7 +1764,7 @@ Hacl_P256_ecdsa_verif_p256_sha512(
 {
   uint64_t m_q[4U] = { 0U };
   uint8_t mHash[64U] = { 0U };
-  Hacl_Hash_SHA2_hash_512(msg, msg_len, mHash);
+  Hacl_Streaming_SHA2_hash_512(msg, msg_len, mHash);
   uint8_t *mHash32 = mHash;
   bn_from_bytes_be4(m_q, mHash32);
   qmod_short(m_q, m_q);
