+++ src/Hacl_RSAPSS.c
@@ -96,17 +96,17 @@ hash(Spec_Hash_Definitions_hash_alg a, uint8_t *mHash, uint32_t msgLen, uint8_t 
   {
     case Spec_Hash_Definitions_SHA2_256:
       {
-        Hacl_Hash_SHA2_hash_256(msg, msgLen, mHash);
+        Hacl_Streaming_SHA2_hash_256(msg, msgLen, mHash);
         break;
       }
     case Spec_Hash_Definitions_SHA2_384:
       {
-        Hacl_Hash_SHA2_hash_384(msg, msgLen, mHash);
+        Hacl_Streaming_SHA2_hash_384(msg, msgLen, mHash);
         break;
       }
     case Spec_Hash_Definitions_SHA2_512:
       {
-        Hacl_Hash_SHA2_hash_512(msg, msgLen, mHash);
+        Hacl_Streaming_SHA2_hash_512(msg, msgLen, mHash);
         break;
       }
     default:
