+++ src/Hacl_SHA2_Vec256.c
@@ -26,8 +26,8 @@
 #include "Hacl_SHA2_Vec256.h"
 
 #include "internal/Hacl_SHA2_Types.h"
-#include "internal/Hacl_SHA2_Generic.h"
 #include "internal/Hacl_Krmllib.h"
+#include "internal/Hacl_Hash_SHA2.h"
 #include "libintvector.h"
 
 static inline void sha224_init8(Lib_IntVector_Intrinsics_vec256 *hash)
