+++ src/Hacl_Streaming_Blake2.c
@@ -63,7 +63,7 @@ void Hacl_Streaming_Blake2_blake2s_32_no_key_init(Hacl_Streaming_Blake2_blake2s_
 /**
   Update function when there is no key; 0 = success, 1 = max length exceeded
 */
-uint32_t
+Hacl_Streaming_Types_error_code
 Hacl_Streaming_Blake2_blake2s_32_no_key_update(
   Hacl_Streaming_Blake2_blake2s_32_state *p,
   uint8_t *data,
@@ -74,7 +74,7 @@ Hacl_Streaming_Blake2_blake2s_32_no_key_update(
   uint64_t total_len = s1.total_len;
   if ((uint64_t)len > (uint64_t)0xffffffffffffffffU - total_len)
   {
-    return (uint32_t)1U;
+    return Hacl_Streaming_Types_MaximumLengthExceeded;
   }
   uint32_t sz;
   if (total_len % (uint64_t)(uint32_t)64U == (uint64_t)0U && total_len > (uint64_t)0U)
@@ -253,7 +253,7 @@ Hacl_Streaming_Blake2_blake2s_32_no_key_update(
         }
       );
   }
-  return (uint32_t)0U;
+  return Hacl_Streaming_Types_Success;
 }
 
 /**
@@ -362,7 +362,7 @@ void Hacl_Streaming_Blake2_blake2b_32_no_key_init(Hacl_Streaming_Blake2_blake2b_
 /**
   Update function when there is no key; 0 = success, 1 = max length exceeded
 */
-uint32_t
+Hacl_Streaming_Types_error_code
 Hacl_Streaming_Blake2_blake2b_32_no_key_update(
   Hacl_Streaming_Blake2_blake2b_32_state *p,
   uint8_t *data,
@@ -373,7 +373,7 @@ Hacl_Streaming_Blake2_blake2b_32_no_key_update(
   uint64_t total_len = s1.total_len;
   if ((uint64_t)len > (uint64_t)0xffffffffffffffffU - total_len)
   {
-    return (uint32_t)1U;
+    return Hacl_Streaming_Types_MaximumLengthExceeded;
   }
   uint32_t sz;
   if (total_len % (uint64_t)(uint32_t)128U == (uint64_t)0U && total_len > (uint64_t)0U)
@@ -572,7 +572,7 @@ Hacl_Streaming_Blake2_blake2b_32_no_key_update(
         }
       );
   }
-  return (uint32_t)0U;
+  return Hacl_Streaming_Types_Success;
 }
 
 /**
