+++ src/Hacl_Streaming_Blake2b_256.c
@@ -75,7 +75,7 @@ Hacl_Streaming_Blake2b_256_blake2b_256_no_key_init(
 /**
   Update function when there is no key; 0 = success, 1 = max length exceeded
 */
-uint32_t
+Hacl_Streaming_Types_error_code
 Hacl_Streaming_Blake2b_256_blake2b_256_no_key_update(
   Hacl_Streaming_Blake2b_256_blake2b_256_state *p,
   uint8_t *data,
@@ -86,7 +86,7 @@ Hacl_Streaming_Blake2b_256_blake2b_256_no_key_update(
   uint64_t total_len = s.total_len;
   if ((uint64_t)len > (uint64_t)0xffffffffffffffffU - total_len)
   {
-    return (uint32_t)1U;
+    return Hacl_Streaming_Types_MaximumLengthExceeded;
   }
   uint32_t sz;
   if (total_len % (uint64_t)(uint32_t)128U == (uint64_t)0U && total_len > (uint64_t)0U)
@@ -285,7 +285,7 @@ Hacl_Streaming_Blake2b_256_blake2b_256_no_key_update(
         }
       );
   }
-  return (uint32_t)0U;
+  return Hacl_Streaming_Types_Success;
 }
 
 /**
