+++ src/Hacl_Streaming_Poly1305_256.c
@@ -70,7 +70,7 @@ Hacl_Streaming_Poly1305_256_init(uint8_t *k, Hacl_Streaming_Poly1305_256_poly130
 /**
 0 = success, 1 = max length exceeded
 */
-uint32_t
+Hacl_Streaming_Types_error_code
 Hacl_Streaming_Poly1305_256_update(
   Hacl_Streaming_Poly1305_256_poly1305_256_state *p,
   uint8_t *data,
@@ -81,7 +81,7 @@ Hacl_Streaming_Poly1305_256_update(
   uint64_t total_len = s.total_len;
   if ((uint64_t)len > (uint64_t)0xffffffffU - total_len)
   {
-    return (uint32_t)1U;
+    return Hacl_Streaming_Types_MaximumLengthExceeded;
   }
   uint32_t sz;
   if (total_len % (uint64_t)(uint32_t)64U == (uint64_t)0U && total_len > (uint64_t)0U)
@@ -254,7 +254,7 @@ Hacl_Streaming_Poly1305_256_update(
         }
       );
   }
-  return (uint32_t)0U;
+  return Hacl_Streaming_Types_Success;
 }
 
 void
